Implemented crud

+++ controllers/feed.js
@@ -1,16 +1,32 @@
 const validator = require('express-validator');
-const Post = require('../models/post');
+
 const fileHelper = require('../util/file');
 
+const Post = require('../models/post');
+const User = require('../models/user');
+
 exports.getPosts = (req, res, next) => {
+    const currentPage = req.query.page || 1;
+    const perPage = 2;
+    let totalItems;
     Post
         .find()
+        .countDocuments()
+        .then(count => {
+            totalItems = count;
+            return Post
+                .find()
+                .skip((currentPage - 1) * perPage)
+                .limit(perPage);
+        })
         .then(posts => {
+            console.log(posts);
             res
                 .status(200)
                 .json({
                     message: 'Fetched posts successfully',
-                    posts: posts
+                    posts: posts,
+                    totalItems: totalItems
                 })
         })
         .catch(err => {
@@ -18,7 +34,7 @@ exports.getPosts = (req, res, next) => {
                 err.statusCode = 500;
             }
             next(err);
-        })
+        });
 }
 
 exports.createPost = (req, res, next) => {
@@ -41,6 +57,7 @@ exports.createPost = (req, res, next) => {
     const content = req.body.content;
     const imageAssetId = req.image.asset_id;
     const imagePublicId = req.image.public_id;
+    let creator;
     
     const post = new Post({
         title: title, 
@@ -48,16 +65,32 @@ exports.createPost = (req, res, next) => {
         imageUrl: imageUrl,
         imageAssetId: imageAssetId,
         imagePublicId: imagePublicId,
-        creator: { name: 'Eric' },
+        creator: req.userId
     });
     post
         .save()
+        .then(result => {
+            return User
+                .findById(req.userId);
+        })
+        .then(user => {
+            creator = user;
+            user
+                .posts
+                .push(post);
+            return user
+                .save();
+        })
         .then(result => {
             res
                 .status(201)
                 .json({
                     message: 'Post created successfully!',
-                    post: result
+                    post: post,
+                    creator: {
+                        id: creator._id,
+                        name: creator.name
+                    }
                 })
         })
         .catch(err => {
@@ -135,6 +168,12 @@ exports.updatePost = (req, res, next) => {
                 throw err;
             }
 
+            if(post.creator.toString() !== req.userId) {
+                const error = new Error('Not authorized');
+                error.statusCode = 403;
+                throw error;
+            }
+
             if(imageUrl !== post.imageUrl) {
                 deleteImage(post.imagePublicId)
             }
@@ -162,6 +201,52 @@ exports.updatePost = (req, res, next) => {
             }
             next(err);
         })
+}
+
+exports.deletePost = (req, res, next) => {
+    const postId = req.params.postId;
+
+    Post
+        .findById({
+            _id: postId
+        })
+        .then(post => {
+            // check logged in user
+            if(!post) {
+                const err = new Error('Could not find post!');
+                err.statusCode = 404;
+                throw err;
+            }
+
+            if(post.creator.toString() !== req.userId) {
+                const error = new Error('Not authorized');
+                error.statusCode = 403;
+                throw error;
+            }
+
+            deleteImage(post.imagePublicId);
+            return Post.findByIdAndRemove(postId);
+        })
+        .then(result => {
+            console.log(result);
+             return User
+                .findById(req.userId);
+        })
+        .then(user => {
+            user.posts.pull(postId);
+            return user.save();
+        })
+        .then(result => {
+            res.status(200).json({
+                message: 'Post deleted!'
+            });
+        })
+        .catch(err => {
+            if(!err.statusCode) {
+                err.statusCode = 500;
+            }
+            next(err);
+        });
 }
 
 const deleteImage = (imagePublicId) => {
